# Optimize Data.Graph.dfs