@treeowl 2023-01-17
Do we have adequate tests for this function? I haven't looked. The benchmarks are *way* too small. That's not your fault, but a benchmark that runs in a few milliseconds doesn't tell us squat. Could you pump the sizes up till we can see something meaningful?