@treeowl 2023-01-17
Other than those points, I love the simplification.