@meooow25 2023-01-19
> That's not your fault...

Technically it is, since I added those benchmarks 😄 

Added a large graph. But I don't understand your position. A large case is useful of course, but surely seeing something decrease for 10ms to 2ms is also meaningful?