@treeowl 2023-01-19
The smaller the time, the less reliable it's likely to be.