@treeowl 2023-01-19
We should really have tests at multiple sizes. I think we do that for sets and maps.