@meooow25 2023-01-19
We do have property tests for graph functions.