@treeowl 2023-01-19
Any guess why there seem to be some regressions? So those go away with even bigger sizes?